Requeue in progress jobs and clean stale lock info on stop
In case there is a failover with a Redis Sentinel cluster with data loss
there can be stale lock information which can cause job processing to be
stuck if the max concurrency limit is reached. Therefore, re-enqueue the
jobs which were in progress and also clean the stale lock info for the
worker pool.

For the cleanup to be thorough, the stop would need to be called on each
worker pool instance.